renaming movebase dotBase
fix imports
