+++ src/main/java/ca/uhn/fhir/jpa/starter/JpaRestfulServer.java
@@ -1,9 +1,10 @@
 package ca.uhn.fhir.jpa.starter;
 
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.jpa.starter.moveBase.PatientInterceptor;
-import ca.uhn.fhir.jpa.starter.moveBase.PatientResponseInterceptor;
-import ca.uhn.fhir.jpa.starter.moveBase.PlainSystemProviderR4;
+import ca.uhn.fhir.jpa.starter.dotBase.PatientInterceptor;
+import ca.uhn.fhir.jpa.starter.dotBase.PatientResponseInterceptor;
+import ca.uhn.fhir.jpa.starter.dotBase.PlainSystemProviderR4;
+
 import javax.servlet.ServletException;
 
 public class JpaRestfulServer extends BaseJpaRestfulServer {
