+++ src/main/java/ca/uhn/fhir/jpa/starter/dotBase/PatientResponseInterceptor.java
@@ -1,6 +1,4 @@
-package ca.uhn.fhir.jpa.starter.moveBase;
-
-import static org.apache.commons.lang3.StringUtils.isNotBlank;
+package ca.uhn.fhir.jpa.starter.dotBase;
 
 /*-
  * #%L
@@ -23,22 +21,15 @@ import static org.apache.commons.lang3.StringUtils.isNotBlank;
  */
 
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.fhirpath.FhirPathExecutionException;
-import ca.uhn.fhir.fhirpath.IFhirPath;
 import ca.uhn.fhir.interceptor.api.Hook;
 import ca.uhn.fhir.interceptor.api.Pointcut;
-import ca.uhn.fhir.rest.api.Constants;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
 import ca.uhn.fhir.rest.api.server.ResponseDetails;
-import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
 import ca.uhn.fhir.util.BundleUtil;
-import ca.uhn.fhir.util.ParametersUtil;
 import java.util.List;
 import org.apache.commons.lang3.tuple.MutablePair;
 import org.apache.commons.lang3.tuple.Pair;
-import org.hl7.fhir.instance.model.api.IBase;
 import org.hl7.fhir.instance.model.api.IBaseBundle;
-import org.hl7.fhir.instance.model.api.IBaseParameters;
 import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.Bundle.BundleEntryComponent;
